fix(picker): make locale property reactive
