+++ packages/oruga/src/components/datetimepicker/types.ts
@@ -1,0 +1,4 @@
+import type { ComponentProps } from "vue-component-type-helpers";
+import Datetimepicker from "./Datetimepicker.vue";
+
+export type DatetimepickerProps = ComponentProps<typeof Datetimepicker>;
